+++ Question service/Question-service/src/main/java/com/dhanashri/Question/service/Module/Response.java
@@ -6,6 +6,6 @@ import lombok.RequiredArgsConstructor;
 @Data
 @RequiredArgsConstructor
 public class Response {
-    private Integer quiz_question_id;
+    private Long quiz_question_id;
     private String response;
 }
